selftests/bpf: Move tracing_multi_bench_attach to bench
In CI, the tracing_multi_bench_attach test could cost >10s, which should
be avoided in daily CI running.

 WATCHDOG: test case tracing_multi_bench_attach executes for 10 seconds...
 kernel-patches#546     tracing_multi_bench_attach:OK

Since it is for benchmark attachment of tracing_multi link, move it to
the generic bench framework.
